break out robinhood types
